feat: likeCount 순으로 정렬하여 post 조회 구현

+++ src/main/java/io/sunflower/like/repository/LikeRepository.java
@@ -9,7 +9,7 @@ public interface LikeRepository extends JpaRepository<Like, Long> {
 
     boolean existsByPostIdAndUserId(Long postId, Long userId);
 
-    Long countByPostId(Long postId);
+    int countByPostId(Long postId);
 
     Optional<Like> findByPostIdAndUserId(Long postId, Long userId);
 }
